perf(analysis): speedup scan
Before this commit, all files in a tree are scanned. This will
use up too much cpu time on unneeded files. This commit introduce
a way to scan only changed files relative to parent commit.
In a case where the file actually deleted from the tree, just skip
it. It's safe because there should be another commit that add
or modifies the file. If there's no parent commit, scan all the
files in that commit tree.